+++ packages/angular-query-experimental/src/devtools/types.ts
@@ -47,13 +47,15 @@ export interface DevtoolsOptions {
   /**
    * The position of the TanStack logo to open and close the devtools panel.
    * `top-left` | `top-right` | `bottom-left` | `bottom-right` | `relative`
-   * Defaults to `bottom-right`.
+   *
+   * @defaultValue bottom-right
    */
   buttonPosition?: DevtoolsButtonPosition
   /**
    * The position of the Angular Query devtools panel.
    * `top` | `bottom` | `left` | `right`
-   * Defaults to `bottom`.
+   *
+   * @defaultValue bottom
    */
   position?: DevtoolsPosition
   /**
@@ -78,7 +80,8 @@ export interface DevtoolsOptions {
   hideDisabledQueries?: boolean
   /**
    * Set this to 'light', 'dark', or 'system' to change the theme of the devtools panel.
-   * Defaults to 'system'.
+   *
+   * @defaultValue system
    */
   theme?: Theme