+++ packages/react-query/src/HydrationBoundary.tsx
@@ -139,7 +139,6 @@ export const HydrationBoundary = ({
               (dehydratedQuery.promise &&
                 existingQuery.state.status !== 'pending' &&
                 existingQuery.state.fetchStatus !== 'fetching' &&
-                dehydratedQuery.dehydratedAt !== undefined &&
                 dehydratedQuery.dehydratedAt >
                   existingQuery.state.dataUpdatedAt)